# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "KUPH Backend API",
    "description": "REST API for KUPH Influencer Marketing Platform. Manages agencies, campaigns, influencers, clients, authentication, subscriptions, and team collaboration.",
    "version": "1.0.0",
    "contact": {
      "name": "KUPH Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /auth/login"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "authProvider": { "type": "string", "nullable": true },
          "authProviderId": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "OnboardAgencyDto": {
        "type": "object",
        "required": ["agencyName"],
        "properties": {
          "agencyName": {
            "type": "string",
            "description": "Name of the agency",
            "example": "My Marketing Agency"
          }
        }
      },
      "Agency": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "currentPlanId": { "type": "string", "format": "uuid", "nullable": true },
          "subscriptionExpiresAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "AgencyMember": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "userId": { "type": "string", "format": "uuid" },
          "agencyId": { "type": "string", "format": "uuid" },
          "role": { "$ref": "#/components/schemas/AgencyRole" },
          "user": { "$ref": "#/components/schemas/User" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "AgencyRole": {
        "type": "string",
        "enum": ["admin", "member", "viewer"]
      },
...
      "AcceptInviteDto": {
        "type": "object",
        "required": ["token"],
        "properties": {
          "token": {
            "type": "string",
            "description": "Invite token from email",
            "example": "a1b2c3d4e5f6..."
          }
        }
      },
...
      "OnboardAgencyResponse": {
        "type": "object",
        "properties": {
          "agency": { "$ref": "#/components/schemas/Agency" },
          "member": { "$ref": "#/components/schemas/AgencyMember" }
        }
      },
      "LookupInviteResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "expiresAt": { "type": "string", "format": "date-time" }
        }
      },
      "Campaign": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "agencyId": { "type": "string", "format": "uuid" },
          "clientId": { "type": "string", "format": "uuid", "nullable": true },
          "name": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "status": { "$ref": "#/components/schemas/CampaignStatus" },
          "budget": { "type": "number", "nullable": true },
          "startDate": { "type": "string", "format": "date", "nullable": true },
          "endDate": { "type": "string", "format": "date", "nullable": true },
          "client": { "$ref": "#/components/schemas/Client" },
          "agency": { "$ref": "#/components/schemas/Agency" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "CreateCampaignDto": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "example": "Summer Launch 2026" },
          "description": { "type": "string", "example": "Influencer campaign for new product line" },
          "clientId": { "type": "string", "format": "uuid" },
          "budget": { "type": "number", "example": 50000 },
          "startDate": { "type": "string", "format": "date", "example": "2026-07-01" },
          "endDate": { "type": "string", "format": "date", "example": "2026-09-30" }
        }
      },
      "UpdateCampaignDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "status": { "$ref": "#/components/schemas/CampaignStatus" },
          "budget": { "type": "number" },
          "startDate": { "type": "string", "format": "date" },
          "endDate": { "type": "string", "format": "date" }
        }
      },
      "CampaignStatus": {
        "type": "string",
        "enum": ["draft", "active", "paused", "completed", "archived"]
      },
      "CampaignShortlist": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "campaignId": { "type": "string", "format": "uuid" },
          "influencerId": { "type": "string", "format": "uuid" },
          "status": { "type": "string", "description": "Shortlist status (approved/pending/rejected)" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "ShareShortlistDto": {
        "type": "object",
        "required": ["influencerIds"],
        "properties": {
          "influencerIds": {
            "type": "array",
            "items": { "type": "string", "format": "uuid" },
            "example": ["uuid-1", "uuid-2", "uuid-3"]
          }
        }
      },
      "ShareShortlistResponse": {
        "type": "object",
        "properties": {
          "shared": {
            "type": "integer",
            "description": "Number of influencers newly shared"
          }
        }
      },
      "Deal": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "campaignId": { "type": "string", "format": "uuid" },
          "influencerId": { "type": "string", "format": "uuid" },
          "status": { "$ref": "#/components/schemas/PitchStatus" },
          "workflowStatus": { "type": "string", "description": "Arbitrary workflow status string" },
          "fee": { "type": "number", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "PitchStatus": {
        "type": "string",
        "enum": ["pending", "approved", "rejected", "negotiating"]
      },
      "UpdateDealStatusDto": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": { "type": "string", "example": "approved" }
        }
      },
      "Client": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "agencyId": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "customAttributes": {
            "type": "object",
            "additionalProperties": true,
            "description": "Flexible key-value metadata"
          },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Influencer": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "agencyId": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "customAttributes": {
            "type": "object",
            "additionalProperties": true,
            "description": "Flexible key-value metadata (e.g. followers, niche, rates)"
          },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "ImportExecuteDto": {
        "type": "object",
        "required": ["entityType", "data"],
        "properties": {
          "entityType": {
            "$ref": "#/components/schemas/CrmEntityType"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": { "type": "string" },
                "customAttributes": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            },
            "description": "Array of entities to import"
          },
          "mappingConfig": {
            "type": "object",
            "additionalProperties": true,
            "description": "Optional column-to-field mapping configuration"
          }
        }
      },
      "ImportExecuteResponse": {
        "type": "object",
        "properties": {
          "imported": {
            "type": "integer",
            "description": "Number of entities successfully imported"
          }
        }
      },
      "CrmEntityType": {
        "type": "string",
        "enum": ["influencer", "brand"]
      },
...
    }
  },
  "paths": {
...
    "/agency/onboard": {
      "post": {
        "tags": ["Agency"],
        "summary": "Create a new agency and set the current user as admin",
        "operationId": "onboardAgency",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/OnboardAgencyDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Agency created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OnboardAgencyResponse" }
              }
            }
          },
          "404": { "description": "User not found" }
        }
      }
    },
...
    "/agency/invite/lookup": {
      "get": {
        "tags": ["Agency"],
        "summary": "Lookup an invite by token (public, no auth required)",
        "operationId": "lookupInvite",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": { "type": "string" },
            "description": "Invite token from email"
          }
        ],
        "responses": {
          "200": {
            "description": "Invite found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LookupInviteResponse" }
              }
            }
          },
          "404": { "description": "Invite not found or expired" }
        }
      }
    },
    "/agency/invite/accept": {
      "post": {
        "tags": ["Agency"],
        "summary": "Accept an invite to join an agency",
        "operationId": "acceptInvite",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AcceptInviteDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully joined agency",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AgencyMember" }
              }
            }
          },
          "404": { "description": "Invite not found" },
          "409": { "description": "User is already a member" }
        }
      }
    },
...
    "/crm/influencers": {
      "get": {
        "tags": ["CRM"],
        "summary": "List all influencers in the agency",
        "operationId": "getInfluencers",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Influencers list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Influencer" }
                }
              }
            }
          }
        }
      }
    },
    "/crm/clients": {
      "get": {
        "tags": ["CRM"],
        "summary": "List all clients (brands) in the agency",
        "operationId": "getClients",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Clients list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Client" }
                }
              }
            }
          }
        }
      }
    },
    "/crm/import/execute": {
      "post": {
        "tags": ["CRM"],
        "summary": "Bulk import influencers or brands from CSV/upload data",
        "operationId": "importExecute",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ImportExecuteDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Import completed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ImportExecuteResponse" }
              }
            }
          },
          "400": { "description": "Unsupported entity type" }
        }
      }
    },
    "/campaigns": {
      "get": {
        "tags": ["Campaigns"],
        "summary": "List all campaigns for the agency",
        "operationId": "getCampaigns",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Campaigns list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Campaign" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Campaigns"],
        "summary": "Create a new campaign",
        "operationId": "createCampaign",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateCampaignDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Campaign created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Campaign" }
              }
            }
          }
        }
      }
    },
    "/campaigns/{id}": {
      "get": {
        "tags": ["Campaigns"],
        "summary": "Get a campaign by ID",
        "operationId": "getCampaign",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Campaign details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Campaign" }
              }
            }
          },
          "404": { "description": "Campaign not found" }
        }
      },
      "patch": {
        "tags": ["Campaigns"],
        "summary": "Update a campaign",
        "operationId": "updateCampaign",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateCampaignDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Campaign updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Campaign" }
              }
            }
          },
          "404": { "description": "Campaign not found" }
        }
      },
      "delete": {
        "tags": ["Campaigns"],
        "summary": "Delete a campaign",
        "operationId": "deleteCampaign",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "204": { "description": "Campaign deleted" },
          "404": { "description": "Campaign not found" }
        }
      }
    },
    "/campaigns/{id}/shortlist": {
      "post": {
        "tags": ["Campaigns"],
        "summary": "Share influencer shortlist with brand for approval",
        "operationId": "shareShortlist",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "Campaign ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ShareShortlistDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Shortlist shared",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ShareShortlistResponse" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Campaigns"],
        "summary": "Get shortlisted influencers for a campaign",
        "operationId": "getShortlist",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Shortlist entries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/CampaignShortlist" }
                }
              }
            }
          }
        }
      }
    },
    "/campaigns/{campaignId}/deals/{dealId}/status": {
      "patch": {
        "tags": ["Campaigns"],
        "summary": "Update workflow status of a deal (e.g. mark as approved, rejected)",
        "operationId": "updateDealStatus",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "campaignId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "dealId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateDealStatusDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Deal status updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Deal" }
              }
            }
          },
          "404": { "description": "Deal not found" }
        }
      }
    },
...
  },
...
}